+++ src/containers/Map/Layers/ControlLayer/CitybikeButton/CitybikeButton.tsx
@@ -1,5 +1,6 @@
 import styles from './CitybikeButton.module.css';
 import btnBgSrc from 'assets/trm.png';
+import { Button } from 'components/UI/Button/Button';
 
 interface CitybikeButton {
   onClick: () => void;
@@ -7,8 +8,8 @@ interface CitybikeButton {
 
 export const CitybikeButton = ({ onClick }: CitybikeButton): JSX.Element => {
   return (
-    <div className={ styles.button } onClick={ onClick }>
+    <Button onClick={ onClick } className={styles.button}>
       <img className={ styles.image } src={ btnBgSrc } alt="Pokaż stacje Torvelo" />
-    </div>
+    </Button>
   );
 }